Python/SudokuSolver/requirements.txt
alembic==1.15.2
attrs==24.1.0
Authlib==1.4.1
autocommand==2.2.2
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.19.0
azure-mgmt-core==1.5.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.2.0
azure-mgmt-subscription==3.1.1
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
boto3==1.36.26
botocore==1.36.26
Brotli==1.1.0
btrfsutil==6.14
build==1.2.2
cached-property==2.0.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==45.0.3
dbus-python==1.4.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
evdev==1.9.0
filelock==3.18.0
Flask==3.1.1
flask-babel==4.0.0
Flask-Compress==1.17
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.5.2
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
freezegun==1.5.1
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
greenlet==3.2.1
h11==0.16.0
httplib2==0.22.0
hyprshade==3.2.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
iniconfig==2.1.0
installer==0.7.0
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
jsonformatter==0.3.4
keyring==25.6.0
ldap3==2.9.1
lensfun==0.3.4
libgravatar==1.0.4
lxml==5.4.0
Mako==1.3.10.dev0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meson==1.8.2
Mirage==0.11.2
moddb==0.12.0
more-itertools==10.7.0
msal==1.32.2
msal-extensions==1.3.1
msrest==0.7.1
numpy==2.3.0
oauthlib==3.2.2
packaging==25.0
paramiko==3.5.1
passlib==1.7.4
pgadmin4==9.2
pillow==11.2.1
platformdirs==4.3.6
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.30.2
psutil==6.1.1
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.22
Pygments==2.19.1
PyGObject==3.52.3
PyJWT==2.10.1
pylspci==0.4.3
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
pytest==8.4.0
python-constraint==1.4.0
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
pywal==3.3.0
PyYAML==6.0.2
qrcode==8.1
Reflector==2023.6.28.0.36.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.11.3
screeninfo==0.8.1
SecretStorage==3.3.3
setuptools==75.9.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.6
speaklater3==1.4
SQLAlchemy==2.0.40
sqlparse==0.5.3
sshtunnel==0.4.0
TBB==0.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.13.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
waypaper==2.6
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
WTForms==3.2.1
zstandard==0.23.0
